ai-service/requirements.txt
accelerate==1.11.0
bitsandbytes==0.48.1
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
einops==0.8.1
filelock==3.19.1
Flask==3.1.2
fsspec==2025.9.0
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
packaging==25.0
pillow==11.3.0
psutil==7.1.1
psycopg2-binary==2.9.11
python-dotenv==1.1.1
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
sentencepiece==0.2.1
sympy==1.13.1
tokenizers==0.22.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.57.1
triton==3.1.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
